Extract functionality into header and make it generic
