+++ depthai_ros_driver/src/ros_xlink.cpp
@@ -52,62 +52,6 @@ public:
 class DepthAI : public dai::DeviceBase {
     using Base = dai::DeviceBase;
 
-    static auto getAllInputs(const dai::Pipeline& pipeline) {
-        const auto& xlinkouts = filterNodesByType<dai::node::XLinkOut>(pipeline);
-        // hack for Output
-        using OutVec = decltype(xlinkouts[0]->getOutputs());
-        using Output = typename OutVec::value_type;
-        struct ret {
-            std::shared_ptr<const dai::node::XLinkOut> node_to;
-            std::vector<Output> out_from;
-        };
-
-        std::vector<ret> out;
-        for (const auto& node : xlinkouts) {
-            // for all input slots of the node
-            ret data;
-            data.node_to = node;
-            const auto& inputs = node->getInputs();
-            for (const auto& inp : inputs) {
-                const auto& links = getInputOf(node, inp.name);
-                // for all links to the slot
-                for (const auto& link : links) {
-                    data.out_from.push_back(link);
-                }
-            }
-            out.emplace_back(std::move(data));
-        }
-        return out;
-    }
-
-    static auto getAllOutputs(const dai::Pipeline& pipeline) {
-        const auto& xlinkins = filterNodesByType<dai::node::XLinkIn>(pipeline);
-        // hack for Input
-        using InVec = decltype(xlinkins[0]->getInputs());
-        using Input = typename InVec::value_type;
-        struct ret {
-            std::shared_ptr<const dai::node::XLinkIn> node_from;
-            std::vector<Input> in_to;
-        };
-
-        std::vector<ret> in;
-        for (const auto& node : xlinkins) {
-            // for all output slots of the node
-            ret data;
-            data.node_from = node;
-            const auto& outputs = node->getOutputs();
-            for (const auto& outp : outputs) {
-                const auto& links = getOutputOf(node, outp.name);
-                // for all links to the slot
-                for (const auto& link : links) {
-                    data.in_to.push_back(link);
-                }
-            }
-            in.emplace_back(std::move(data));
-        }
-        return in;
-    }
-
 public:
     using Base::Base;
 
@@ -122,6 +66,7 @@ public:
         Base::closeImpl();
     }
 
+protected:
     // create stream based on name at the time of subscription
     template <class MsgType>
     auto generate_cb_lambda(std::string name) {
@@ -168,35 +113,6 @@ public:
         return Base::startPipelineImpl(pipeline);
     }
 
-    /**
-     * @tparam Range vector of Inputs or Outputs
-     */
-    template <class Range>
-    dai::DatatypeEnum _getCommonType(const Range& links) {
-        std::optional<dai::DatatypeEnum> common_type;
-        for (const auto& link : links) {
-            // get the link type
-            const auto& allowed_types = link.possibleDatatypes;
-            const dai::DatatypeEnum type = [&] {
-                if (allowed_types.size() > 1) {
-                    return dai::DatatypeEnum::Buffer;
-                } else {
-                    return allowed_types[0].datatype;
-                }
-            }();
-            if (!common_type) {
-                common_type = type;
-            } else if (common_type.value() != type) {
-                // @TODO: find the common type, not necessarily the buffer
-                common_type = dai::DatatypeEnum::Buffer;
-            }
-        }
-        if (!common_type) {
-            return dai::DatatypeEnum::Buffer;
-        }
-        return common_type.value();
-    }
-
     void _setup_publishers(const dai::Pipeline& pipeline) {
         // get all XLinkOut
         const auto& out_links = getAllInputs(pipeline);
@@ -205,7 +121,7 @@ public:
             // get name for publisher
             const auto& name = node->getStreamName();
 
-            auto common_type = _getCommonType(node_links.out_from);
+            auto common_type = getCommonType(node_links.out_from);
 
             // create appropriate publisher using the common_type
             switch (common_type) {
@@ -238,6 +154,7 @@ public:
             }
         }
     }
+
     void _setup_subscribers(const dai::Pipeline& pipeline) {
         // get all XLinkIn
         const auto& in_links = getAllOutputs(pipeline);
@@ -246,7 +163,7 @@ public:
             // get name for publisher
             const auto& name = node->getStreamName();
 
-            auto common_type = _getCommonType(node_links.in_to);
+            auto common_type = getCommonType(node_links.in_to);
 
             // create appropriate subscriber using the common_type
             switch (common_type) {
